perbaikan bug renew customer

+++ database/migrations/2024_05_23_164015_create_payments_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('subscription_id');
             $table->unsignedBigInteger('customer_id');
-            $table->integer('amount');
-            $table->integer('fee')->default(0);
+            $table->decimal('amount',10,2);
+            $table->decimal('fee',10,2)->default(0);
             $table->enum('status', ['paid', 'unpaid', 'pending'])->default('pending');
             $table->date('tanggal_bayar')->nullable();
             $table->string('checkout_link')->nullable();
